feat: Implement Inventory report logic, UI cleanup, and animations

+++ CampusEats.Backend/Common/DTOs/AuthDto.cs
@@ -29,7 +29,7 @@ public class UserDto
     public string Role { get; set; } = string.Empty;
     public string PhoneNumber { get; set; } = string.Empty;
     public string? StudentId { get; set; }
-    public int LoyaltyPoints { get; set; }
+    public decimal LoyaltyPoints { get; set; }
     public bool IsActive { get; set; }
     public DateTime CreatedAt { get; set; }
 }
