+++ CampusEats.Backend/Common/DTOs/LoyaltyDto.cs
@@ -3,9 +3,9 @@
 public record LoyaltyPointsDto
 {
     public Guid UserId { get; init; }
-    public int CurrentPoints { get; init; }
-    public int TotalEarned { get; init; }
-    public int TotalRedeemed { get; init; }
+    public decimal CurrentPoints { get; init; }
+    public decimal TotalEarned { get; init; }
+    public decimal TotalRedeemed { get; init; }
     public decimal PointsValue { get; init; }
 }
 
@@ -13,7 +13,7 @@ public record LoyaltyTransactionDto
 {
     public Guid Id { get; init; }
     public Guid UserId { get; init; }
-    public int PointsChange { get; init; }
+    public decimal PointsChange { get; init; }
     public string Type { get; init; } = string.Empty;
     public string Description { get; init; } = string.Empty;
     public Guid? OrderId { get; init; }
@@ -22,8 +22,8 @@ public record LoyaltyTransactionDto
 
 public record RedeemPointsResultDto
 {
-    public int PointsRedeemed { get; init; }
+    public decimal PointsRedeemed { get; init; }
     public decimal DiscountAmount { get; init; }
-    public int RemainingPoints { get; init; }
+    public decimal RemainingPoints { get; init; }
     public string Message { get; init; } = string.Empty;
 }
