+++ CampusEats.Backend/Common/DTOs/PaymentDto.cs
@@ -20,5 +20,5 @@ public class PaymentDto
     public DateTime? PaidAt { get; set; }
     public DateTime? FailedAt { get; set; }
     public string? FailureReason { get; set; }
-    public int? LoyaltyPointsUsed { get; set; }
+    public decimal? LoyaltyPointsUsed { get; set; }
 }
