+++ CampusEats.Backend/Domain/User.cs
@@ -9,7 +9,7 @@ public class User
     public string Role { get; set; } = string.Empty;
     public string PhoneNumber { get; set; } = string.Empty;
     public string? StudentId { get; set; }
-    public int LoyaltyPoints { get; set; } = 0;
+    public decimal LoyaltyPoints { get; set; } = 0;
     public bool IsActive { get; set; } = true; 
     
     public DateTime CreatedAt { get; set; }
